PostgreSQL, Use prepared statements for queries with and without binds, based on the usage count of the given SQL string.
Currently all queries without binds are not cached as prepared statement.
However all queries with binds are cached as prepared statement on it's
first use. Using prepared statements based on the usage count of a given
query, reduces network round trips for preparation and deallocation of
single-use queries. Most notably it allowes creation of prepared statements
of repeated queries without binds.

The CountedStatementPool is optimized for performance and low memory footprint.
This is a very simple hash implementation based on an Array.
It takes 8 bytes per entry instead of around 50 of a Hash based one on 64-bit MRI.
It depends on String#hash, so that collisions are expected and can occur.
But they have a minor performance impact only, so they should be no real issue.

There are two new configuration variables, that are hard coded for now:
The size of the CountedStatementPool (8192) and the usage count of a query
to be moved into the PreparedStatementPool (2). These could be added as another
connection parameter.

This patch also tunes PreparedStatementPool insofar, that queries are removed
in LRU order, when the pool is full, which is a little bit fairer/faster. And
the deallocation is done in bulks of each 5% of the pool size, in order to save
database roundtrips.