PostgreSQL, Use prepared statements for queries with and without binds, based on the usage count of the given SQL string.
This introduces a second statement pool, which is used for each new query
string. Any query that is used twice is moved to the prepared statement pool.

When the pool limit for prepared queries is reached, then queries are
removed in LRU order, now. Deallocation is done in bulks of each 5% of the
pool size to save DB roundtrips.