PostgreSQL, Add input type casts for primitive types.
Ruby-pg's default way to serialize values for transmission to the database is to
call #to_s . This however creates a temporary String object for each value.
Setting a class based type map avoids the allocation of this addional String.

The performace benefit is measureable in microbenchmarks, but not with
the overhead of activerecord. However it's free to use and has no drawback.