PostgreSQL: Use a specialized Result object to defer retrieval of values
PG::Result stores values received from the database internally in a compact
memory format. When the value is accessed, it is converted to a ruby string or
casted to another ruby object.

While ActiveRecord::Result enforces the conversion of all rows and columns to
ruby objects, this implementation defers conversion to when the single value is
accessed. Therefore ActiveRecord::ConnectionAdapters::PostgreSQL::Result,
DeferredRow and PG::Tuple store a reference to the PG::Result object.

DeferredRow and PG::Tuple cache values after being materialized. They detach the
PG::Result object, when all columns have been materialized or when the object is
marshaled.